fix: resolve 2 import layer violations (routes_rescore, _file_lock)
- routes_rescore.py: import read_run_data from services.ports instead of data.fs
- _file_lock.py: move canonical implementation from adapters/ to data/,
  update analysis/subagents re-export to use data._file_lock
