+++ src/quodeq/analysis/subagents/_file_lock.py
@@ -1,2 +1,2 @@
-"""Re-export — moved to quodeq.adapters._file_lock (infrastructure layer)."""
-from quodeq.adapters._file_lock import lock_file, unlock_file  # noqa: F401
+"""Re-export — canonical implementation in quodeq.data._file_lock."""
+from quodeq.data._file_lock import lock_file, unlock_file  # noqa: F401
